% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\alias{ptriang}
\alias{qtriang}
\alias{rtriang}
\title{Triangular Distribution}
\usage{
dtriang(x, min = 0, max = 1, mode = 0.5)

ptriang(q, min = 0, max = 1, mode = 0.5)

qtriang(p, min = 0, max = 1, mode = 0.5)

rtriang(n, min = 0, max = 1, mode = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{min}{lower limit (a).}

\item{max}{upper limit (b).}

\item{mode}{mode (c).}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution, quantile and random generation for the triangular distribution.
}
